unsigned intに変更した箇所をsize_tに修正

+++ DreamSat/Camera.cydsn/communicate_to_camera.c
@@ -29,14 +29,14 @@ typedef enum {
 }RecievePhaseCode;
 
 //Variables
-static Byte*            camera_buff;
-static size_t           camera_buff_length;
-static unsigned long    recv_count;
-RecievePhaseCode        recv_phase;
+static Byte*        camera_buff;
+static size_t       camera_buff_length;
+static size_t       recv_count;
+RecievePhaseCode    recv_phase;
 
 void IsrCamRx()
 {
-    unsigned long index = recv_count;
+    size_t index = recv_count;
     Byte recv_data = UART_TO_CAMERA_GetChar();
